Refactored CLI code
