+++ src/main/java/net/sf/jabref/gui/remote/JabRefMessageHandler.java
@@ -16,29 +16,24 @@
 package net.sf.jabref.gui.remote;
 
 import net.sf.jabref.JabRef;
+import net.sf.jabref.cli.ArgumentProcessor;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.logic.remote.server.MessageHandler;
 
-import java.util.Optional;
-import java.util.Vector;
+import java.util.List;
 
 public class JabRefMessageHandler implements MessageHandler {
 
-    private final JabRef jabRef;
-
-    public JabRefMessageHandler(JabRef jabRef) {
-        this.jabRef = jabRef;
-    }
-
     @Override
     public void handleMessage(String message) {
-        Optional<Vector<ParserResult>> loaded = jabRef.processArguments(message.split("\n"), false);
-        if (!(loaded.isPresent())) {
+        ArgumentProcessor argumentProcessor = new ArgumentProcessor(message.split("\n"), false);
+        if (!(argumentProcessor.hasParserResults())) {
             throw new IllegalStateException("Could not start JabRef with arguments " + message);
         }
 
-        for (int i = 0; i < loaded.get().size(); i++) {
-            ParserResult pr = loaded.get().elementAt(i);
+        List<ParserResult> loaded = argumentProcessor.getParserResults().get();
+        for (int i = 0; i < loaded.size(); i++) {
+            ParserResult pr = loaded.get(i);
             JabRef.mainFrame.addParserResult(pr, i == 0);
         }
     }
